refactor: split command data into files
