+++ src/bot/commands/Information/user.ts
@@ -1,4 +1,4 @@
-import { ApplicationCommandOptionType, ApplicationCommandType, PermissionFlagsBits } from 'discord.js';
+import { PermissionFlagsBits } from 'discord.js';
 import { Command, CommandRunOptions } from '../../../structures/Command';
 import type { DenkyClient } from '../../../types/Client';
 
@@ -14,61 +14,6 @@ export default class UserCommand extends Command {
       guildOnly: false,
     };
     this.permissions = { bot: [PermissionFlagsBits.EmbedLinks], user: [] };
-
-    this.addRawOptions({
-      name: 'user',
-      type: ApplicationCommandType.ChatInput,
-      description: client.languages.manager.get('en_US', 'commandDescriptions:user'),
-      descriptionLocalizations: {
-        'pt-BR': client.languages.manager.get('pt_BR', 'commandDescriptions:user'),
-      },
-      options: [
-        {
-          name: 'info',
-          type: ApplicationCommandOptionType.Subcommand,
-          description: client.languages.manager.get('en_US', 'commandDescriptions:user/info'),
-          descriptionLocalizations: {
-            'pt-BR': client.languages.manager.get('pt_BR', 'commandDescriptions:user/info'),
-          },
-          options: [
-            {
-              name: client.languages.manager.get('en_US', 'commandNames:user/info/user'),
-              nameLocalizations: {
-                'pt-BR': client.languages.manager.get('pt_BR', 'commandNames:user/info/user'),
-              },
-              type: ApplicationCommandOptionType.User,
-              required: false,
-              description: client.languages.manager.get('en_US', 'commandDescriptions:user/info/user'),
-              descriptionLocalizations: {
-                'pt-BR': client.languages.manager.get('pt_BR', 'commandDescriptions:user/info/user'),
-              },
-            },
-          ],
-        },
-        {
-          name: 'avatar',
-          type: ApplicationCommandOptionType.Subcommand,
-          description: client.languages.manager.get('en_US', 'commandDescriptions:user/avatar'),
-          descriptionLocalizations: {
-            'pt-BR': client.languages.manager.get('pt_BR', 'commandDescriptions:user/avatar'),
-          },
-          options: [
-            {
-              name: client.languages.manager.get('en_US', 'commandNames:user/avatar/user'),
-              nameLocalizations: {
-                'pt-BR': client.languages.manager.get('pt_BR', 'commandNames:user/avatar/user'),
-              },
-              type: ApplicationCommandOptionType.User,
-              required: false,
-              description: client.languages.manager.get('en_US', 'commandDescriptions:user/avatar/user'),
-              descriptionLocalizations: {
-                'pt-BR': client.languages.manager.get('pt_BR', 'commandDescriptions:user/avatar/user'),
-              },
-            },
-          ],
-        },
-      ],
-    });
   }
 
   override run({ t, interaction }: CommandRunOptions) {
