+++ src/bot/commands/Moderation/poll.ts
@@ -1,8 +1,6 @@
-import { ApplicationCommandOptionType, ApplicationCommandSubCommandData, ApplicationCommandType, EmbedBuilder, Message, PermissionFlagsBits } from 'discord.js';
+import { EmbedBuilder, Message, PermissionFlagsBits } from 'discord.js';
 import { Command, CommandRunOptions } from '../../../structures/Command';
 import type { DenkyClient } from '../../../types/Client';
-
-type PollAcceptableOptions = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9;
 
 const MappedNumericEmojis = {
   1: '1️⃣',
@@ -28,45 +26,6 @@ export default class PollCommand extends Command {
       guildOnly: true,
     };
     this.permissions = { bot: [PermissionFlagsBits.EmbedLinks], user: [PermissionFlagsBits.ManageMessages] };
-
-    const opts: ApplicationCommandSubCommandData['options'] = [];
-    for (let i: PollAcceptableOptions = 1; i <= 9; (i as PollAcceptableOptions)++) {
-      opts.push({
-        name: client.languages.manager.get('en_US', `commandNames:poll/create/option${i}`),
-        nameLocalizations: {
-          'pt-BR': client.languages.manager.get('pt_BR', `commandNames:poll/create/option${i}`),
-        },
-        type: ApplicationCommandOptionType.String,
-        required: i === 1,
-        description: client.languages.manager.get('en_US', `commandDescriptions:poll/create/option${i}`),
-        descriptionLocalizations: {
-          'pt-BR': client.languages.manager.get('pt_BR', `commandDescriptions:poll/create/option${i}`),
-        },
-      });
-    }
-
-    this.addRawOptions({
-      name: client.languages.manager.get('en_US', 'commandNames:poll'),
-      nameLocalizations: {
-        'pt-BR': client.languages.manager.get('pt_BR', 'commandNames:poll'),
-      },
-      type: ApplicationCommandType.ChatInput,
-      description: client.languages.manager.get('en_US', 'commandNames:poll'),
-      options: [
-        {
-          name: client.languages.manager.get('en_US', 'commandNames:poll/create'),
-          nameLocalizations: {
-            'pt-BR': client.languages.manager.get('pt_BR', 'commandNames:poll/create'),
-          },
-          type: ApplicationCommandOptionType.Subcommand,
-          description: client.languages.manager.get('en_US', 'commandDescriptions:poll/create'),
-          descriptionLocalizations: {
-            'pt-BR': client.languages.manager.get('pt_BR', 'commandDescriptions:poll/create'),
-          },
-          options: opts,
-        },
-      ],
-    });
   }
 
   override async run({ t, interaction }: CommandRunOptions) {
