+++ src/bot/commands/Text/text.ts
@@ -1,4 +1,3 @@
-import { ApplicationCommandOptionType, ApplicationCommandSubCommandData, ApplicationCommandType } from 'discord.js';
 import { Command, CommandRunOptions } from '../../../structures/Command';
 import type { DenkyClient } from '../../../types/Client';
 
@@ -14,77 +13,6 @@ export default class TextCommand extends Command {
       guildOnly: false,
     };
     this.permissions = { bot: [], user: [] };
-
-    const baseTextOption: ApplicationCommandSubCommandData['options'] = [
-      {
-        name: client.languages.manager.get('en_US', 'commandNames:text'),
-        nameLocalizations: {
-          'pt-BR': client.languages.manager.get('pt_BR', 'commandNames:text'),
-        },
-        type: ApplicationCommandOptionType.String,
-        required: true,
-        description: client.languages.manager.get('en_US', 'commandDescriptions:text/text'),
-        descriptionLocalizations: {
-          'pt-BR': client.languages.manager.get('pt_BR', 'commandDescriptions:text/text'),
-        },
-      },
-    ];
-
-    this.addRawOptions({
-      name: client.languages.manager.get('en_US', 'commandNames:text'),
-      nameLocalizations: {
-        'pt-BR': client.languages.manager.get('pt_BR', 'commandNames:text'),
-      },
-      type: ApplicationCommandType.ChatInput,
-      description: client.languages.manager.get('en_US', 'commandDescriptions:text'),
-      descriptionLocalizations: {
-        'pt-BR': client.languages.manager.get('pt_BR', 'commandDescriptions:text'),
-      },
-      options: [
-        {
-          name: client.languages.manager.get('en_US', 'commandNames:text/claps'),
-          nameLocalizations: {
-            'pt-BR': client.languages.manager.get('pt_BR', 'commandNames:text/claps'),
-          },
-          type: ApplicationCommandOptionType.Subcommand,
-          description: client.languages.manager.get('en_US', 'commandDescriptions:text/claps'),
-          descriptionLocalizations: {
-            'pt-BR': client.languages.manager.get('pt_BR', 'commandDescriptions:text/claps'),
-          },
-          options: baseTextOption,
-        },
-        {
-          name: 'emojify',
-          type: ApplicationCommandOptionType.Subcommand,
-          description: client.languages.manager.get('en_US', 'commandDescriptions:text/emojify'),
-          descriptionLocalizations: {
-            'pt-BR': client.languages.manager.get('pt_BR', 'commandDescriptions:text/emojify'),
-          },
-          options: baseTextOption,
-        },
-        {
-          name: client.languages.manager.get('en_US', 'commandNames:text/invert'),
-          nameLocalizations: {
-            'pt-BR': client.languages.manager.get('pt_BR', 'commandNames:text/invert'),
-          },
-          type: ApplicationCommandOptionType.Subcommand,
-          description: client.languages.manager.get('en_US', 'commandDescriptions:text/invert'),
-          descriptionLocalizations: {
-            'pt-BR': client.languages.manager.get('pt_BR', 'commandDescriptions:text/invert'),
-          },
-          options: baseTextOption,
-        },
-        {
-          name: 'vaporwave',
-          type: ApplicationCommandOptionType.Subcommand,
-          description: client.languages.manager.get('en_US', 'commandDescriptions:text/vaporwave'),
-          descriptionLocalizations: {
-            'pt-BR': client.languages.manager.get('pt_BR', 'commandDescriptions:text/vaporwave'),
-          },
-          options: baseTextOption,
-        },
-      ],
-    });
   }
 
   override run({ t, interaction }: CommandRunOptions) {
