+++ src/bot/commands/Utils/data/help.ts
@@ -1,0 +1,20 @@
+import { ApplicationCommandType } from 'discord.js';
+import { CommandDataStructure } from '../../../../structures/CommandDataStructure';
+import type { DenkyClient } from '../../../../types/Client';
+
+export default class HelpData extends CommandDataStructure {
+  constructor(client: DenkyClient) {
+    super(client);
+    this.data = {
+      name: client.languages.manager.get('en_US', 'commandNames:help'),
+      nameLocalizations: {
+        'pt-BR': client.languages.manager.get('pt_BR', 'commandNames:help'),
+      },
+      type: ApplicationCommandType.ChatInput,
+      description: client.languages.manager.get('en_US', 'commandDescriptions:help'),
+      descriptionLocalizations: {
+        'pt-BR': client.languages.manager.get('pt_BR', 'commandDescriptions:help'),
+      },
+    };
+  }
+}
