+++ src/bot/commands/Utils/data/ping.ts
@@ -1,0 +1,17 @@
+import { ApplicationCommandType } from 'discord.js';
+import { CommandDataStructure } from '../../../../structures/CommandDataStructure';
+import type { DenkyClient } from '../../../../types/Client';
+
+export default class PingData extends CommandDataStructure {
+  constructor(client: DenkyClient) {
+    super(client);
+    this.data = {
+      name: 'ping',
+      type: ApplicationCommandType.ChatInput,
+      description: client.languages.manager.get('en_US', 'commandDescriptions:ping'),
+      descriptionLocalizations: {
+        'pt-BR': client.languages.manager.get('pt_BR', 'commandDescriptions:ping'),
+      },
+    };
+  }
+}
